Creation of the Tikz-GeomEdit Connector
Functional push of some Tikz code on the geometric constructions as well
as the pull from them.
The Tikz code editor can be opened using the contextual menu (right
click) and choosing "Open Tikz Editor".
Added two README files that explain the system architecture and the
current scope of application.